requirements.txt
streamlit==1.31.0
streamlit-webrtc==0.47.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
aiortc==1.5.0
